Move authenticators
